+++ mizani/breaks.py
@@ -33,7 +33,7 @@ if typing.TYPE_CHECKING:
     from mizani.typing import (
         DatetimeBreaksUnits,
         DurationUnit,
-        FloatVector,
+        FloatArrayLike,
         NDArrayFloat,
         NDArrayTimedelta,
         Timedelta,
@@ -81,7 +81,7 @@ class breaks_log:
         self.n = n
         self.base = base
 
-    def __call__(self, limits: TupleFloat2) -> FloatVector:
+    def __call__(self, limits: TupleFloat2) -> NDArrayFloat:
         """
         Compute breaks
 
@@ -149,7 +149,7 @@ class _breaks_log_sub:
         self.n = n
         self.base = base
 
-    def __call__(self, limits: TupleFloat2) -> FloatVector:
+    def __call__(self, limits: TupleFloat2) -> NDArrayFloat:
         base = self.base
         n = self.n
         rng = log(limits, base)
@@ -209,6 +209,9 @@ class minor_breaks:
     """
     Compute minor breaks
 
+    This is the naive method. It does not take into account
+    the transformation.
+
     Parameters
     ----------
     n : int
@@ -237,10 +240,10 @@ class minor_breaks:
 
     def __call__(
         self,
-        major: FloatVector,
+        major: FloatArrayLike,
         limits: Optional[TupleFloat2] = None,
         n: Optional[int] = None,
-    ) -> FloatVector:
+    ) -> NDArrayFloat:
         """
         Minor breaks
 
@@ -265,7 +268,9 @@ class minor_breaks:
             return np.array([])
 
         if limits is None:
-            limits = min_max(major)
+            low, high = min_max(major)
+        else:
+            low, high = min_max(limits)
 
         if n is None:
             n = self.n
@@ -285,7 +290,7 @@ class minor_breaks:
             mbreaks.append(lhs + factors * sep)
 
         minor = np.hstack(mbreaks)
-        minor = minor.compress((limits[0] <= minor) & (minor <= limits[1]))
+        minor = minor.compress((low <= minor) & (minor <= high))
         return minor
 
 
@@ -312,19 +317,22 @@ class minor_breaks_trans:
     >>> from mizani.transforms import sqrt_trans
     >>> major = [1, 2, 3, 4]
     >>> limits = [0, 5]
-    >>> sqrt_trans().minor_breaks(major, limits)
-    array([0.5, 1.5, 2.5, 3.5, 4.5])
-    >>> class sqrt_trans2(sqrt_trans):
-    ...     def __init__(self):
-    ...         self.minor_breaks = minor_breaks_trans(sqrt_trans2)
-    >>> sqrt_trans2().minor_breaks(major, limits)
+    >>> t1 = sqrt_trans()
+    >>> t1.minor_breaks(major, limits)
     array([1.58113883, 2.54950976, 3.53553391])
+
+    # Changing the regular `minor_breaks` method
+
+    >>> t2 = sqrt_trans()
+    >>> t2.minor_breaks = minor_breaks()
+    >>> t2.minor_breaks(major, limits)
+    array([0.5, 1.5, 2.5, 3.5, 4.5])
 
     More than 1 minor break
 
     >>> major = [1, 10]
     >>> limits = [1, 10]
-    >>> sqrt_trans().minor_breaks(major, limits, 4)
+    >>> t2.minor_breaks(major, limits, 4)
     array([2.8, 4.6, 6.4, 8.2])
     """
 
@@ -334,10 +342,10 @@ class minor_breaks_trans:
 
     def __call__(
         self,
-        major: FloatVector,
+        major: FloatArrayLike,
         limits: Optional[TupleFloat2] = None,
         n: Optional[int] = None,
-    ) -> FloatVector:
+    ) -> NDArrayFloat:
         """
         Minor breaks for transformed scales
 
@@ -358,12 +366,6 @@ class minor_breaks_trans:
         out : array_like
             Minor breaks
         """
-        if not self.trans.dataspace_is_numerical:
-            raise TypeError(
-                "trans_minor_breaks can only be used for data "
-                "whose format is numerical."
-            )
-
         if limits is None:
             limits = min_max(major)
 
@@ -372,11 +374,11 @@ class minor_breaks_trans:
 
         major = self._extend_breaks(major)
         major = self.trans.inverse(major)
-        limits = self.trans.inverse(limits)  # type: ignore
+        limits = self.trans.inverse(limits)
         minor = minor_breaks(n)(major, limits)
         return self.trans.transform(minor)
 
-    def _extend_breaks(self, major: FloatVector) -> FloatVector:
+    def _extend_breaks(self, major: FloatArrayLike) -> FloatArrayLike:
         """
         Append 2 extra breaks at either end of major
 
@@ -499,7 +501,7 @@ class breaks_timedelta:
     [0.0, 5.0, 10.0, 15.0, 20.0, 25.0]
     """
 
-    _calculate_breaks: Callable[[TupleFloat2], FloatVector]
+    _calculate_breaks: Callable[[TupleFloat2], NDArrayFloat]
 
     def __init__(self, n: int = 5, Q: Sequence[float] = (1, 2, 5, 10)):
         self._calculate_breaks = breaks_extended(n=n, Q=Q)
@@ -655,7 +657,7 @@ class timedelta_helper:
         """
         _min = self.limits[0] / self.factor
         _max = self.limits[1] / self.factor
-        return _min, _max  # type: ignore
+        return _min, _max
 
     def timedelta_to_numeric(
         self, timedeltas: NDArrayTimedelta
@@ -665,7 +667,7 @@ class timedelta_helper:
         """
         return np.array([self.to_numeric(td) for td in timedeltas])
 
-    def numeric_to_timedelta(self, values: FloatVector) -> NDArrayTimedelta:
+    def numeric_to_timedelta(self, values: NDArrayFloat) -> NDArrayTimedelta:
         """
         Convert sequence of numerical values to timedelta
         """
@@ -804,7 +806,7 @@ class breaks_extended:
         # a score. Return 1 ignores all that.
         return 1
 
-    def __call__(self, limits: TupleFloat2) -> FloatVector:
+    def __call__(self, limits: TupleFloat2) -> NDArrayFloat:
         """
         Calculate the breaks